// src/services/rest-api/customer-service/model/address-info-model.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Customer Service API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 0.0.267
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */



/**
 * 
 * @export
 * @interface AddressInfoModel
 */
export interface AddressInfoModel {
    /**
     * 
     * @type {string}
     * @memberof AddressInfoModel
     */
    'addressTypeCode'?: string | null;
    /**
     * 
     * @type {string}
     * @memberof AddressInfoModel
     */
    'addressTypeDesc'?: string | null;
    /**
     * 
     * @type {string}
     * @memberof AddressInfoModel
     */
    'addressNo'?: string | null;
    /**
     * 
     * @type {string}
     * @memberof AddressInfoModel
     */
    'buildingOrVillage'?: string | null;
    /**
     * 
     * @type {string}
     * @memberof AddressInfoModel
     */
    'floor'?: string | null;
    /**
     * 
     * @type {string}
     * @memberof AddressInfoModel
     */
    'roomNo'?: string | null;
    /**
     * 
     * @type {string}
     * @memberof AddressInfoModel
     */
    'moo'?: string | null;
    /**
     * 
     * @type {string}
     * @memberof AddressInfoModel
     */
    'soi'?: string | null;
    /**
     * 
     * @type {string}
     * @memberof AddressInfoModel
     */
    'street'?: string | null;
    /**
     * 
     * @type {string}
     * @memberof AddressInfoModel
     */
    'provinceCode'?: string | null;
    /**
     * 
     * @type {string}
     * @memberof AddressInfoModel
     */
    'provinceNameTh'?: string | null;
    /**
     * 
     * @type {string}
     * @memberof AddressInfoModel
     */
    'districtCode'?: string | null;
    /**
     * 
     * @type {string}
     * @memberof AddressInfoModel
     */
    'districtNameTh'?: string | null;
    /**
     * 
     * @type {string}
     * @memberof AddressInfoModel
     */
    'subDistrictCode'?: string | null;
    /**
     * 
     * @type {string}
     * @memberof AddressInfoModel
     */
    'subDistrictNameTh'?: string | null;
    /**
     * 
     * @type {string}
     * @memberof AddressInfoModel
     */
    'countryCode'?: string | null;
    /**
     * 
     * @type {string}
     * @memberof AddressInfoModel
     */
    'countryDesc'?: string | null;
    /**
     * 
     * @type {string}
     * @memberof AddressInfoModel
     */
    'zipCode'?: string | null;
    /**
     * 
     * @type {string}
     * @memberof AddressInfoModel
     */
    'customAddress1'?: string | null;
    /**
     * 
     * @type {string}
     * @memberof AddressInfoModel
     */
    'customAddress2'?: string | null;
    /**
     * 
     * @type {string}
     * @memberof AddressInfoModel
     */
    'customAddress3'?: string | null;
    /**
     * 
     * @type {string}
     * @memberof AddressInfoModel
     */
    'customAddress4'?: string | null;
    /**
     * 
     * @type {string}
     * @memberof AddressInfoModel
     */
    'telephoneNo'?: string | null;
    /**
     * 
     * @type {string}
     * @memberof AddressInfoModel
     */
    'mobileNo'?: string | null;
    /**
     * 
     * @type {string}
     * @memberof AddressInfoModel
     */
    'officeNo'?: string | null;
    /**
     * 
     * @type {string}
     * @memberof AddressInfoModel
     */
    'email'?: string | null;
    /**
     * 
     * @type {string}
     * @memberof AddressInfoModel
     */
    'docReceiveChannel'?: string | null;
    /**
     * 
     * @type {string}
     * @memberof AddressInfoModel
     */
    'docReceiveAddressType'?: string | null;
}